# Phase 1: optimization