--- a/lib/app/modules/home/views/child_home_view.dart
+++ b/lib/app/modules/home/views/child_home_view.dart
@@ -19,6 +19,9 @@ class ChildHomeView extends GetView<ChildHomeController> {
 
   @override
   Widget build(BuildContext context) {
+    // Set up navigation using helper
+    controller.setupNavigation();
+    
     // Debug logging (can be removed in production)
     AdaptiveSizes.logDeviceInfo(context);