--- a/lib/widgets/progress_node_widget.dart
+++ b/lib/widgets/progress_node_widget.dart
@@ -1,6 +1,4 @@
 import 'package:flutter/material.dart';
-import 'package:get/get.dart';
-import 'package:kioku_navi/utils/sizes.dart';
 import 'package:kioku_navi/utils/adaptive_sizes.dart';
 import 'package:kioku_navi/widgets/rounded_button.dart';
 
@@ -67,35 +65,25 @@ class ProgressNodeWidget extends StatelessWidget {
   }) : size = size ?? 18.0;
 
   /// Gets the adaptive stroke width based on node size
-  static double getAdaptiveStrokeWidth(double nodeSize) {
-    final context = Get.context;
-    if (context != null) {
-      return AdaptiveSizes.getProgressStrokeWidth(context);
-    }
-    // Fallback for when context is not available
-    return 6.0;
+  static double getAdaptiveStrokeWidth(BuildContext context) {
+    return AdaptiveSizes.getProgressStrokeWidth(context);
   }
 
   /// Gets the adaptive padding between progress and button based on node size
-  static double getAdaptivePadding(double nodeSize) {
-    final context = Get.context;
-    if (context != null) {
-      return AdaptiveSizes.getProgressPadding(context, nodeSize);
-    }
-    // Fallback for when context is not available
-    return nodeSize * 0.03;
+  static double getAdaptivePadding(BuildContext context, double nodeSize) {
+    return AdaptiveSizes.getProgressPadding(context, nodeSize);
   }
 
   @override
   Widget build(BuildContext context) {
     // If there's active progress, show progress indicator around the button
     if (state == NodeState.active && completionPercentage < 100.0) {
-      final double adaptivePadding = getAdaptivePadding(size);
+      final double adaptivePadding = getAdaptivePadding(context, size);
 
       return Stack(
         alignment: Alignment.center,
         children: [
-          _buildActiveNodeProgress(),
+          _buildActiveNodeProgress(context),
           Padding(
             padding: EdgeInsets.all(adaptivePadding), // Adaptive padding
             child: _buildRoundedButton(),
@@ -209,16 +197,16 @@ class ProgressNodeWidget extends StatelessWidget {
   }
 
   /// Creates a circular progress indicator for nodes with incomplete progress.
-  Widget _buildActiveNodeProgress() {
+  Widget _buildActiveNodeProgress(BuildContext context) {
     // Calculate progress size to wrap around the button + padding
-    final double adaptivePadding = getAdaptivePadding(size);
+    final double adaptivePadding = getAdaptivePadding(context, size);
     final double paddingSize = adaptivePadding * 2; // padding on both sides
     final double progressSize =
         size + paddingSize + (size * 0.15); // extra space for progress ring
     final double progress = completionPercentage / 100.0;
 
     // Get adaptive stroke width based on node size
-    final double strokeWidth = getAdaptiveStrokeWidth(size);
+    final double strokeWidth = getAdaptiveStrokeWidth(context);
 
     return SizedBox(
       width: progressSize,