… OPTIMIZATION_REPORT.md
… PHASE_1_OPTIMIZATION_REPORT.md
… lib/widgets/course_section_widget.dart